ClosedPulls: extract the component
Let's keep the main app component clean.
