+++ frontend/src/pulldasher/index.tsx
@@ -5,9 +5,9 @@ import { Column } from '../column';
 import { QACompare } from './sort';
 import { LeaderList, getLeaders } from '../leader-list';
 import { useMyPullNotification, useMyReviewNotification } from "./notifications"
-import { useStyleConfig, Flex, Spacer, Box, SimpleGrid, VStack } from "@chakra-ui/react"
-import { ClosedPullCard } from "../pull-card";
+import { Box, SimpleGrid, VStack } from "@chakra-ui/react"
 import { useBoolUrlState } from "../use-url-state";
+import { ClosedPulls } from '../closed-pulls';
 
 export const Pulldasher: React.FC = function() {
    const allPulls = useAllPulls();
@@ -51,21 +51,4 @@ export const Pulldasher: React.FC = function() {
          {showClosedPulls && <ClosedPulls onClickClose={toggleShowClosedPulls}/>}
       </Box>
    </>);
-}
-
-function ClosedPulls({onClickClose}: {onClickClose: () => void}) {
-   const closedPulls: Pull[] = Array.from(useAllPulls()).filter(pull => pull.closed_at);
-   const styles = useStyleConfig('Column', {variant: "closed"});
-   return (
-      <Box __css={styles} position="absolute" right="0" top="70" bottom="0" width="300px" boxShadow="0px 0px 10px 0px #00000020">
-         <Flex className="column_header" onClick={onClickClose}>
-            <Box p={3} pl={4}>Recently Closed Pulls</Box>
-            <Spacer/>
-            <Box className="pull_count" p={3}>{closedPulls.length}</Box>
-         </Flex>
-         {closedPulls.map(pull =>
-            <ClosedPullCard key={Math.random()} pull={pull} show={true}/>
-         )}
-      </Box>
-   );
 }
